requirements.txt
Django>=5.2
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
django-cors-headers
